Add context to applications
The context serves as an interface with the application and its
environment. It stores the bookmark and, depending on the application,
success flags indicating the action that should be performed after the
x11 app has run.
Add a default constructor for DateTime in order to allow for easy
creation of bookmarks and consequently context.