Add text widget
Also add a dirty usage of the text widget to showcase its functionality.
There was an issue when adding sevaral lines by pressing Enter, since
the Application took Enter also as a command to proceed to the next
Dialog. Consequently, change the command to proceed to the next Dialog
to Tab, which also feels more natural.
Also add Shift+Tab functionality for moving backward to prior dialogs.

Also note a TODO in order to fix a surfaced bug.