Defer changing the bookmark prior to saving
Be able to revise querystring when going back in the dialogs.

Prior, if the querystring was set to be removed and the next dialog was
opened, the querystring would be cut off from the url. When going back,
the url would lack a querystring to remove or to keep.

Fix this by storing the decision whether to keep the querystring in the
app context and act accordingly prior to saving the url, rather than
changing the url in the Dialog.

Also move a function to align it with the declaration orde in the header
file.