Merge add-settings and fetch-settings
Merge the settings in order to allow for more resilient settings.
If the settings file misses an individual setting, the apps take a
default now, instead of failing.
I decided to use a default setting structure. It is a single source of
truth with regard to the defaults, both when it comes to writing
settings file and when picking individual default settings. This is the
reason I chose to merge the settings for bbm-add and bbm-fetch. If I had
taken them separately, writing a settings file with defaults would need
to take both settings, for add and fetch, into account. Thus, the
either defaults would need to be available in either application,
anyway. Keeping them split would just add more complexity compared to
the little benefit it provides. The merge on the other hand also reduces
overall program complexity.